requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-cors-headers==4.6.0
django-cuser==2017.3.16
django-debug-toolbar==5.1.0
django-environ==0.11.2
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-modeltranslation==0.19.9
django-mptt==0.16.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
magic-filter==1.0.12
modeltranslation==0.25
multidict==6.1.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
schedule==1.2.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.14.0
